Update README.md
Hi @EionRobb, thanks for this pidgin-ignore-nickchange .. very useful! When I couldn't see the plugin in Pidgin after building, it took me a moment of jumping into the source to figure that the name of this plugin (under which it shows up in the Plugins list) is actually "Nick Change Hiding", and not "Ignore Nick Change" as listed on https://developer.pidgin.im/wiki/ThirdPartyPlugins#Notifications - perhaps you'd like to clarify that in this README? And/or get them to correct this on the pidgin wiki. Best & thanks again for this.